Moved cudf::table to its own file. Made types.hpp into simple type declarations.
