+++ cpp/tests/table/table_tests.cu
@@ -16,7 +16,7 @@
 
 #include <tests/utilities/cudf_test_fixtures.h>
 #include <tests/utilities/column_wrapper.cuh>
-#include <types.hpp>
+#include <table.hpp>
 #include <utilities/type_dispatcher.hpp>
 
 #include <gmock/gmock.h>
